Fix the filename computation of article list

+++ test/e2e/bm.e2e.test.ts
@@ -15,7 +15,7 @@ const parameters = {
   format: ['nopic'],
 }
 
-await testAllRenders(parameters, async (outFiles) => {
+await testAllRenders('bm-wikipedia', parameters, async (outFiles) => {
   test(`test zim integrity for ${outFiles[0]?.renderer} renderer`, async () => {
     await expect(zimcheck(outFiles[0].outFile)).resolves.not.toThrowError()
   })
@@ -29,8 +29,8 @@ await testAllRenders(parameters, async (outFiles) => {
       if (dump.nopic) {
         // nopic has enough files (this is just an estimate and can change
         // with time, as new Mediwiki versions are released).
-        expect(dump.status.files.success).toBeGreaterThan(12)
-        expect(dump.status.files.success).toBeLessThan(22)
+        expect(dump.status.files.success).toBeGreaterThan(6)
+        expect(dump.status.files.success).toBeLessThan(14)
         // nopic has enough redirects
         expect(dump.status.redirects.written).toBeGreaterThan(170)
         // nopic has enough articles
@@ -40,11 +40,13 @@ await testAllRenders(parameters, async (outFiles) => {
   })
 
   afterAll(() => {
-    rimraf.sync(`./${outFiles[0].testId}`)
+    if (!process.env.KEEP_ZIMS) {
+      rimraf.sync(`./${outFiles[0].testId}`)
+    }
   })
 })
 
-await testAllRenders({ ...parameters, addNamespaces: 1 }, async (outFiles) => {
+await testAllRenders('bm-wikipedia-with-ns-1', { ...parameters, addNamespaces: 1 }, async (outFiles) => {
   test(`Articles with "Discussion" namespace for ${outFiles[0]?.renderer} renderer for bm.wikipedia.org`, async () => {
     await execa('redis-cli flushall', { shell: true })
 
@@ -55,6 +57,8 @@ await testAllRenders({ ...parameters, addNamespaces: 1 }, async (outFiles) => {
     expect(discussionArticlesList.length).toBeGreaterThan(30)
   })
   afterAll(() => {
-    rimraf.sync(`./${outFiles[0].testId}`)
+    if (!process.env.KEEP_ZIMS) {
+      rimraf.sync(`./${outFiles[0].testId}`)
+    }
   })
 })
