+++ test/e2e/en10.e2e.test.ts
@@ -14,7 +14,7 @@ const parameters = {
   format: ['nopic', 'nopdf'],
 }
 
-await testAllRenders(parameters, async (outFiles) => {
+await testAllRenders('en10-wikipedia', parameters, async (outFiles) => {
   describe('en10', () => {
     test(`Simple articleList for ${outFiles[0]?.renderer} renderer`, async () => {
       // Created 2 outputs
@@ -24,8 +24,8 @@ await testAllRenders(parameters, async (outFiles) => {
         if (dump.nopic) {
           // nopic has enough files (this is just an estimate and can change
           // with time, as new Mediwiki versions are released).
-          expect(dump.status.files.success).toBeGreaterThan(50)
-          expect(dump.status.files.success).toBeLessThan(60)
+          expect(dump.status.files.success).toBeGreaterThanOrEqual(48)
+          expect(dump.status.files.success).toBeLessThan(59)
           // nopic has enough redirects
           expect(dump.status.redirects.written).toBeGreaterThan(500)
           // nopic has 10 articles
@@ -63,7 +63,9 @@ await testAllRenders(parameters, async (outFiles) => {
     })
 
     afterAll(() => {
-      rimraf.sync(`./${outFiles[0].testId}`)
+      if (!process.env.KEEP_ZIMS) {
+        rimraf.sync(`./${outFiles[0].testId}`)
+      }
     })
   })
 })
